requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic>=2.12.0a1
PyYAML==6.0.2
